Split audio/graphics functions from tools.py into own files
