+++ tuxemon/core/event/actions/play_sound.py
@@ -24,6 +24,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+import tuxemon.core.audio
 from tuxemon.core import tools
 from tuxemon.core.db import db
 from tuxemon.core.event.eventaction import EventAction
@@ -41,5 +42,5 @@ class PlaySoundAction(EventAction):
 
     def start(self):
         filename = self.parameters.filename
-        sound = tools.load_sound(filename)
+        sound = tuxemon.core.audio.load_sound(filename)
         sound.play()
