+++ tuxemon/core/event/actions/translated_dialog.py
@@ -28,7 +28,8 @@ import logging
 
 from tuxemon.core.locale import process_translate_text
 from tuxemon.core.event.eventaction import EventAction
-from tuxemon.core.tools import open_dialog, get_avatar
+from tuxemon.core.tools import open_dialog
+from tuxemon.core.graphics import get_avatar
 
 logger = logging.getLogger(__name__)
