+++ tuxemon/core/states/combat/combat.py
@@ -40,6 +40,7 @@ from itertools import chain
 
 import pygame
 
+import tuxemon.core.audio
 from tuxemon.core import state, tools
 from tuxemon.core.combat import check_status, fainted, get_awake_monsters, defeated
 from tuxemon.core.locale import T
@@ -425,11 +426,11 @@ class CombatState(CombatAnimations):
         def add(menuitem):
             monster = menuitem.game_object
             if monster.current_hp == 0:
-                tools.open_dialog(self.game, [T.format("combat_fainted", parameters={"name": monster.name})])
+                graphics.open_dialog(self.game, [T.format("combat_fainted", parameters={"name": monster.name})])
             elif monster in self.active_monsters:
-                tools.open_dialog(self.game, [T.format("combat_isactive", parameters={"name": monster.name})])
+                graphics.open_dialog(self.game, [T.format("combat_isactive", parameters={"name": monster.name})])
                 msg = T.translate("combat_replacement_is_fainted")
-                tools.open_dialog(self.game, [msg])
+                graphics.open_dialog(self.game, [msg])
             else:
                 self.add_monster_into_play(player, monster)
                 self.game.pop_state()
@@ -515,7 +516,7 @@ class CombatState(CombatAnimations):
         x, y, w, h = self.game.screen.get_rect()
         rect = pygame.Rect(0, 0, w, h // 4)
         rect.bottomright = w, h
-        border = tools.load_and_scale(self.borders_filename)
+        border = graphics.load_and_scale(self.borders_filename)
         self.dialog_box = GraphicBox(border, None, self.background_color)
         self.dialog_box.rect = rect
         self.sprites.add(self.dialog_box, layer=100)
@@ -643,7 +644,7 @@ class CombatState(CombatAnimations):
             message = ''
 
         try:
-            tools.load_sound(technique.sfx).play()
+            tuxemon.core.audio.load_sound(technique.sfx).play()
         except AttributeError:
             pass
 
@@ -803,7 +804,7 @@ class CombatState(CombatAnimations):
         frame_time = .09
         images = list()
         for fn in technique.images:
-            image = tools.load_and_scale(fn)
+            image = graphics.load_and_scale(fn)
             images.append((image, frame_time))
 
         tech = PygAnimation(images, False)
