+++ tuxemon/core/states/splash/__init__.py
@@ -34,6 +34,7 @@ from __future__ import unicode_literals
 
 import logging
 
+import tuxemon.core.audio
 from tuxemon.core import prepare
 from tuxemon.core import state
 from tuxemon.core import tools
@@ -68,7 +69,7 @@ class SplashState(state.State):
         cc = self.load_sprite("gfx/ui/intro/creative_commons.png")
         cc.rect.topleft = width - splash_border - cc.rect.width, height - splash_border - cc.rect.height
 
-        self.ding = tools.load_sound("sound_ding")
+        self.ding = tuxemon.core.audio.load_sound("sound_ding")
         self.ding.play()
 
     def process_event(self, event):
