+++ tuxemon/core/ui/__init__.py
@@ -62,7 +62,7 @@ class UserInterface(object):
 
         # Handle loading a single image, multiple images, or surfaces
         if images_type == 'str' or images_type == 'unicode':
-            surface = tools.load_and_scale(images)
+            surface = graphics.load_and_scale(images)
             self.images = [(surface, animation_speed)]
 
         elif images_type == 'list' or images_type == 'tuple':
@@ -72,7 +72,7 @@ class UserInterface(object):
                 item_type = type(item).__name__
 
                 if item_type == 'str' or item_type == 'unicode':
-                    surface = tools.load_and_scale(images)
+                    surface = graphics.load_and_scale(images)
                 else:
                     surface = item
                 self.images.append((surface, animation_speed))
